more adjustments while constantly reexecuting all tests
