+++ CsCore/xUnitTests/src/Plugins/CsCoreXUnitTests/com/csutil/tests/io/FileTests.cs
@@ -253,7 +253,7 @@ namespace com.csutil.tests {
         }
 
         private static DirectoryInfo CreateDirectoryForTesting(string name) {
-            var rootDir = EnvironmentV2.instance.GetRootTempFolder().GetChildDir(name);
+            var rootDir = EnvironmentV2.instance.GetRootTempFolder().GetChildDir("FileTests").GetChildDir(name);
             rootDir.DeleteV2(); // ensure that the test folder does not yet exist
             rootDir.CreateV2();
             Assert.True(rootDir.Exists);
